refactor test utils
